improvements

+++ src/main/java/hr/fer/zemris/bachelor/LCS/Environment/Environment.java
@@ -3,7 +3,7 @@ package hr.fer.zemris.bachelor.LCS.Environment;
 public interface Environment {
 
     boolean[] getInput();
-    double getReward(double action);
-    boolean checkOutput(double output);
+    double getReward(int action);
+    boolean checkOutput(int output);
 
 }
